feat(bookmark button): added bookmark button to guides

+++ src/routes/guides/[id]/+page.server.ts
@@ -1,4 +1,8 @@
-import { deleteGuideSchema, toggleGuideUsefulSchema } from '@/schemas/guide';
+import {
+	deleteGuideSchema,
+	toggleGuideUsefulSchema,
+	toggleGuideBookmarkSchema,
+} from '@/schemas/guide';
 import type { GuideWithAuthor, ModerationInfo } from '@/types/types';
 import { handleFormAction } from '@/utils';
 import { error, fail, redirect } from '@sveltejs/kit';
@@ -62,7 +66,25 @@ export const load = async (event) => {
 		return { count: usefuls.count, userUseful: usefuls.has_useful };
 	}
 
+	async function getBookmark(id: string): Promise<{ userBookmark: boolean }> {
+		const { data: bookmark, error: bookmarkError } = await event.locals.supabase
+			.rpc('get_guide_bookmark', {
+				guide_id: parseInt(id),
+				user_id: user?.id,
+			})
+			.single();
+
+		if (bookmarkError) {
+			const errorMessage = 'Error fetching bookmark, please try again later.';
+			setFlash({ type: 'error', message: errorMessage }, event.cookies);
+			return error(500, errorMessage);
+		}
+
+		return { userBookmark: bookmark.has_bookmark };
+	}
+
 	const usefulCount = await getUsefulCount(event.params.id);
+	const bookmark = await getBookmark(event.params.id);
 
 	return {
 		guide: await getGuide(event.params.id),
@@ -76,6 +98,13 @@ export const load = async (event) => {
 			zod(toggleGuideUsefulSchema),
 			{
 				id: 'toggle-guide-useful',
+			}
+		),
+		toggleBookmarkForm: await superValidate(
+			{ value: bookmark.userBookmark },
+			zod(toggleGuideBookmarkSchema),
+			{
+				id: 'toggle-guide-bookmark',
 			}
 		),
 	};
@@ -120,6 +149,43 @@ export const actions = {
 				} else {
 					const { error: supabaseError } = await event.locals.supabase
 						.from('guides_useful')
+						.delete()
+						.eq('guide_id', parseInt(event.params.id))
+						.eq('user_id', userId);
+
+					if (supabaseError) {
+						setFlash({ type: 'error', message: supabaseError.message }, event.cookies);
+						return fail(500, { message: supabaseError.message, form });
+					}
+				}
+
+				return { form };
+			}
+		),
+
+	toggleBookmark: async (event) =>
+		handleFormAction(
+			event,
+			toggleGuideBookmarkSchema,
+			'toggle-guide-bookmark',
+			async (event, userId, form) => {
+				if (form.data.value) {
+					const { error: supabaseError } = await event.locals.supabase
+						.from('guides_bookmark')
+						.insert([
+							{
+								guide_id: parseInt(event.params.id),
+								user_id: userId,
+							},
+						]);
+
+					if (supabaseError) {
+						setFlash({ type: 'error', message: supabaseError.message }, event.cookies);
+						return fail(500, { message: supabaseError.message, form });
+					}
+				} else {
+					const { error: supabaseError } = await event.locals.supabase
+						.from('guides_bookmark')
 						.delete()
 						.eq('guide_id', parseInt(event.params.id))
 						.eq('user_id', userId);
